Ajustes de otimização de tempo de execução

+++ backend/src/db/Models/vehicle_models/vehicle_new_model.py
@@ -2,7 +2,6 @@ from datetime import datetime
 from typing import Optional
 
 from pydantic import BaseModel, Field
-
 from Utils.Internacionalization import ModelsDescriptionTexts
 
 
@@ -14,6 +13,6 @@ class VehicleNewModel(BaseModel):
     color: str = Field(max_length=25, description=ModelsDescriptionTexts.VEHICLE_COLOR)
     license_plate: str = Field(max_length=10, description=ModelsDescriptionTexts.LICENSE_PLATE)
     cd_fuel_type: int = Field(description=ModelsDescriptionTexts.CD_FUEL_TYPE)
-    fuel_capacity: int = Field(default=1, gt=1, description=ModelsDescriptionTexts.FUEL_CAPACITY)
+    fuel_capacity: float = Field(default=1.0, gt=1.0, description=ModelsDescriptionTexts.FUEL_CAPACITY)
     odometer: Optional[float] = Field(default=0.0, description=ModelsDescriptionTexts.ODOMETER)
     is_active: bool = Field(default=True, description=ModelsDescriptionTexts.IS_ACTIVE)
